+++ src/components/side-bar/aboutus-page-side-bar.js
@@ -1,12 +1,12 @@
 import { SITE_META } from '../about-us/constants/data/index'
 import { buildFbShareLink } from '../about-us/utils/build-fb-share-link'
 import { colors, typography } from '../../themes/common-variables'
-import { donatePath } from '../../constants/index'
 import { replaceGCSUrlOrigin } from '@twreporter/core/lib/utils/storage-url-processor'
 import { screen } from '../about-us/utils/screen'
 import { storageUrlPrefix } from '../about-us/utils/config'
 import anchorlist from '../about-us/constants/data/sidebar-anchor'
 import baseComponents from './base-components'
+import DonationLink from '@twreporter/react-components/lib/donation-link-with-utm'
 import hrefs from '../about-us/constants/data/sidebar-link'
 import PropTypes from 'prop-types'
 import React from 'react'
@@ -129,9 +129,11 @@ class AboutusPageSideBar extends React.PureComponent {
             currentAnchorId={currentAnchorId}
           />
           <Icons>
-            <a href={donatePath} target="_blank">
+            <DonationLink
+              utmMedium="about-us"
+            >
               <img src={`${replaceGCSUrlOrigin(`${storageUrlPrefix}/sidebar-icon1.png`)}`} />
-            </a>
+            </DonationLink>
             <a href={hrefs.subscribe} target="_blank">
               <img src={`${replaceGCSUrlOrigin(`${storageUrlPrefix}/sidebar-icon2.png`)}`} />
             </a>